move akondiselftest to widgets as it uses the widgets lib
core calls the test binary but the binary uses widgets, so we cannot
have it in the core package as it would be circular deping.
instead move the binary to the widgets lib and let core recommend it.
looking at the code the dialog not being there should not make the
core lib non-functional anyway.